refactored test projects to remove dependencies per #830 (#831)
* refactored test projects to remove dependencies per #830

* fix namespace
