+++ tests/Test.ADAL.NET.Common/AssertException.cs
