+++ tests/Test.ADAL.NET.Common/Certs/invalid_cert.pfx
