+++ tests/Test.ADAL.NET.Common/Certs/valid_cert.cer
