+++ tests/Test.ADAL.NET.Common/Certs/valid_cert2.pfx
