+++ tests/Test.ADAL.NET.Common/Log.cs
@@ -1,4 +1,4 @@
-//------------------------------------------------------------------------------
+﻿//----------------------------------------------------------------------
 //
 // Copyright (c) Microsoft Corporation.
 // All rights reserved.
@@ -25,47 +25,15 @@
 //
 //------------------------------------------------------------------------------
 
-using System;
-using System.Collections.Generic;
+using System.Diagnostics;
 
-namespace Test.ADAL.Common
+namespace Test.ADAL.NET.Common
 {
-    public abstract class AdalTestsBase
+    internal static class Log
     {
-        static AdalTestsBase()
+        internal static void Comment(string message, params object[] args)
         {
-            StsDictionary = new Dictionary<StsType, Sts>();
-        }
-
-        public static Dictionary<StsType, Sts> StsDictionary { get; private set; }
-
-        protected Sts Sts { get; set; }
-
-#if !TEST_ADAL_WINRT_UNIT
-        public Microsoft.VisualStudio.TestTools.UnitTesting.TestContext TestContext { get; set; }
-
-        protected StsType GetStsTypeFromContext()
-        {
-            return GetStsType((string)TestContext.DataRow["StsType"]);
-        }
-#endif
-
-        protected static StsType GetStsType(string stsType)
-        {
-            return (StsType)Enum.Parse(typeof(StsType), stsType);
-        }
-
-        protected static Sts SetupStsService(StsType stsType)
-        {
-            Sts sts;
-
-            if (!StsDictionary.TryGetValue(stsType, out sts))
-            {
-                sts = StsFactory.CreateSts(stsType);
-                StsDictionary.Add(stsType, sts);
-            }
-
-            return sts;
+            Debug.WriteLine(message, args);
         }
     }
 }
