+++ tests/Test.ADAL.NET.Common/SerializationHelper.cs
