+++ tests/Test.ADAL.NET.Common/Verify.cs
