+++ tests/Test.ADAL.NET.Common/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NSubstitute" version="2.0.3" targetFramework="net461" />
+</packages>
