+++ tests/Test.ADAL.NET.Unit/CommonUnitTests.cs
@@ -30,7 +30,7 @@ using System;
 using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Test.ADAL.Common
+namespace Test.ADAL.NET.Unit
 {
     public static class CommonUnitTests
     {
