+++ tests/Test.ADAL.NET.Unit/Test.ADAL.NET.Unit.csproj
@@ -1,10 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <!-- Import common settings -->
   <Import Project="..\..\build\ADAL.Net.settings.targets" />
   <PropertyGroup Label="SonarQube analysis settings">
     <!-- These properties have no effect on how the project is built. -->
     <SonarQubeTestProject>true</SonarQubeTestProject>
+    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -14,7 +15,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>Test.ADAL.NET.Unit</RootNamespace>
     <AssemblyName>Test.ADAL.NET.Unit</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
@@ -71,27 +72,15 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="..\Test.ADAL.Common\CommonUnitTests.cs">
-      <Link>Common\CommonUnitTests.cs</Link>
-    </Compile>
-    <Compile Include="..\Test.ADAL.Common\Log.cs">
-      <Link>Common\Log.cs</Link>
-    </Compile>
-    <Compile Include="..\Test.ADAL.Common\TokenCacheTests.cs">
-      <Link>Common\TokenCacheTests.cs</Link>
-    </Compile>
+    <Compile Include="CommonUnitTests.cs" />
+    <Compile Include="TokenCacheTests.cs" />
     <Compile Include="AdalDotNetTests.cs" />
     <Compile Include="AuthenticationParametersTests.cs" />
     <Compile Include="ClaimsChallengeExceptionTests.cs" />
-    <Compile Include="Common\AssertException.cs" />
     <Compile Include="LoggerCallbackTests.cs" />
-    <Compile Include="Mocks\MockHelpers.cs" />
-    <Compile Include="Mocks\MockHttpMessageHandler.cs" />
-    <Compile Include="Mocks\MockWebUI.cs" />
     <Compile Include="NonInteractiveTests.cs" />
     <Compile Include="OboFlowTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestConstants.cs" />
     <Compile Include="TokenCacheKeyTests.cs" />
     <Compile Include="TokenCacheUnitTests.cs" />
     <Compile Include="InstanceDiscoveryTests.cs" />
@@ -112,11 +101,11 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
-    <None Include="..\Test.ADAL.Common\Certs\valid_cert.pfx">
+    <None Include="..\Test.ADAL.NET.Common\Certs\valid_cert.pfx">
       <Link>valid_cert.pfx</Link>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Include="..\Test.ADAL.Common\Certs\valid_cert2.pfx">
+    <None Include="..\Test.ADAL.NET.Common\Certs\valid_cert2.pfx">
       <Link>valid_cert2.pfx</Link>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
@@ -125,6 +114,10 @@
     <ProjectReference Include="..\..\src\Microsoft.IdentityModel.Clients.ActiveDirectory\Microsoft.IdentityModel.Clients.ActiveDirectory.csproj">
       <Project>{aece489d-9ed4-442d-8d8f-0f383aebb781}</Project>
       <Name>Microsoft.IdentityModel.Clients.ActiveDirectory</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Test.ADAL.NET.Common\Test.ADAL.NET.Common.csproj">
+      <Project>{412f8f05-2694-4d2e-8fe0-a821721cf682}</Project>
+      <Name>Test.ADAL.NET.Common</Name>
     </ProjectReference>
   </ItemGroup>
   <Choose>
