+++ tests/Test.ADAL.NET.Unit/TokenCacheTests.cs
@@ -32,6 +32,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Test.ADAL.NET.Common;
 
 namespace Test.ADAL.Common.Unit
 {
