--- tests/Test.ADAL.NET.Unit/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NSubstitute" version="1.10.0.0" targetFramework="net45" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="4.0.1" targetFramework="net45" />
-</packages>
